requirements.txt
python-dotenv
pymongo[srv]
pymongo
certifi
numpy
pandas
matplotlib
seaborn
scikit-learn
catboost
xgboost
dill
flask
pyaml
-e .